Change start input type from float to int

+++ calculator.py
@@ -1,4 +1,4 @@
-start = float(input("Starting number : "))
+start = int(input("Starting number : "))
 percentage = float(input("Percentage number : "))
 years = float(input("Number of years : "))
 def calculation(initial_amount, rate, duration):
@@ -6,6 +6,6 @@ def calculation(initial_amount, rate, duration):
   for i in range(duration):
     total = total + (total * rate / 100)
     pure_gain = total - initial_amount
-    return pure_gain
+  return pure_gain
 final_result = calculation(start, percentage, years)
 print(f"The gain is {final_result:.2f} or {final_result + start:.2f}")
